fix: moved estimate delete to ee
