+++ web/ce/components/estimates/index.ts
@@ -1,3 +1,3 @@
 export * from "./estimate-list-item-buttons";
-
 export * from "./update";
+export * from "./points";
