+++ web/core/components/estimates/points/index.ts
@@ -1,7 +1,6 @@
 export * from "./preview";
 export * from "./create";
 export * from "./update";
-export * from "./delete";
 export * from "./select-dropdown";
 
 export * from "./create-root";
