+++ web/ee/components/estimates/index.ts
@@ -1,2 +1,3 @@
 export * from "./estimate-list-item-buttons";
 export * from "./update";
+export * from "./points";
