+++ web/ee/components/estimates/points/delete.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/estimates/points/delete";
