+++ web/ee/components/estimates/points/index.ts
@@ -1,0 +1,1 @@
+export * from "./delete";
